Wrote some tests for Process/minor internal mods
One currently doesn't pass - it doesn't fail, mind you, but the test
never terminates. A read that is happening in one of the tests is
blocking, but it is reliant on the input of the previous process.

Perhaps, I modify it that if there isn't a line immediately, it
recursively requests a line from the parent process; this isn't neat and
is prone to lockup either way.

I also added a buffer to the stdout too, as some might build up in a
process that is piped to, and not read frequently. Unfortunately, with
that addition comes another problem - if the line isn't read explicitly
(currently only via >>), the functors DO get called, but the buffer is
never empties and is increasingly filled up.

I also wrote a new testing program - it simply out puts a lot of data,
so that we can test the pipe congestion better. I just wrote this one
because we're gonna start writing the processes that are invoked from
the testing program from scratch - enabling the tests to be run
cross-platform.