Procesess with functors can no longer be read from
Using operator>> or .read is prohibited for processes that have
processes/files to output to, or if they have a functor to execute.

The reason is to avoid massive memory increases for those who don't need
the output, and only wish to use a functor.

I need to fix the tests so they compile however. Tomorrow's job!